Fix existing font-end API data expectations
The front end was expecting JSON objects from the server that had
misleading and inefficient properties. I renamed 'name' > to 'userInfo'
and set that to be single object instead of an array of a single object.
Moreover, 'reply' was renamed to 'replies', given that will be an array
of multiple objects. Finally, all instances of '_id' were renamed to 'id' to be more DB
agostic.